+++ CHANGELOG.md
@@ -3,10 +3,36 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 7.4.0 - 2025-04-10
+## 8.2.0 - 2025-04-10
 ### Added
 - Functionality to allow renames in glue event listener. Table will be created with new name, copied partitions 
 from old to new table, and dropped old table.
+
+## 8.1.2 - 2025-04-08
+### Fix
+- Hive version validation when multiple listeners are used.
+
+## 8.1.1 - Release Action Failed, no release made.
+
+## 8.1.0 - 2025-03-28
+### Added
+- Support for MSK Kafka clusters.
+
+## 8.0.2 - 2024-08-29
+### Changed
+- `ApiaryGlueSync` override `skipArchive` to be `true` by default. Backward incompatible behavior that turns off table archiving by default. Allows for per table overrides if needed. If you don't rely on Glue table version it is safe to upgrade to this version of the glue listener without making changes.
+
+## 7.3.11 - 2024-06-19
+### Changed
+- Upgrade aws version from `1.11.520` to `1.12.276` in `apiary-receiver-sqs`.
+
+## 7.3.10 - 2024-06-18
+### Added
+- Add `aws-java-sdk-sts` to enable IRSA authentication in `apiary-receiver-sqs`.
+
+## 7.3.9 - 2024-06-03
+### Added
+- Add `aws-java-sdk-sts` to enable IRSA authentication in `apiary-metastore-listener`.
 
 ## 7.3.8 - 2023-08-02
 ### Added